% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_SMC_iw.R
\name{ABC_SMC_iw}
\alias{ABC_SMC_iw}
\title{ABC_SMC_iw}
\usage{
ABC_SMC_iw(
  obs_data,
  calc_ss_function = calc_error_all,
  init_epsilon_values = init_epsilon,
  prior_generating_function = prior_gen,
  prior_density_function = prior_dens,
  number_of_particles = number_of_particles,
  sigma = 0.05,
  stop_rate = 0.001,
  num_iterations = num_iterations,
  idparsopt = c(1, 2, 3, 4, 5),
  pars = parameters
)
}
\description{
ABC_SMC_iw
}
