% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_log.R
\name{calc_log_lik_DAISIE_iw}
\alias{calc_log_lik_DAISIE_iw}
\title{Calculates the loglikelihood of the DAISIE model with island-wide
diversity-dependence given data and a set of model parameters}
\usage{
calc_log_lik_DAISIE_iw(params, datalist)
}
\value{
a numeric represents the loglikelihood
}
\description{
Calculates the loglikelihood of the DAISIE model with island-wide
diversity-dependence given data and a set of model parameters
}
