% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_log.R
\name{calc_log_prior_DAISIE_iw}
\alias{calc_log_prior_DAISIE_iw}
\title{Calculates the log prior density}
\usage{
calc_log_prior_DAISIE_iw(params, idparsopt)
}
\value{
a numeric represents the log prior density
}
\description{
Calculates the log prior density
}
