% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MLE.R
\name{get_MLE}
\alias{get_MLE}
\title{Maximum-likelihood estimation}
\usage{
get_MLE(the_sim, pars_use)
}
\arguments{
\item{the_sim}{simulation output from DAISIE IW simulation framework}

\item{pars_use}{paramters used to generate the simulation output}
}
\value{
a list
}
\description{
Maximum-likelihood estimation
}
