% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ABC.R
\docType{data}
\name{param_space}
\alias{param_space}
\title{Run ABC-SMC with DAISIE-IW data}
\format{
An object of class \code{data.frame} with 320 rows and 6 columns.
}
\usage{
param_space
}
\arguments{
\item{param_set}{A numeric of a specific line of parameter set in parameter space}

\item{idparsopt}{A vector of positions of the parameters that be optimized.}

\item{number_of_particles}{The number of particles in each iteration.}

\item{num_iterations}{The maximum number of iterations.}

\item{print_frequency}{The frequency of printing out the progress.}

\item{sigma}{The standard deviation of the normal distribution used to generate the proposal distribution.}

\item{stop_rate}{The stopping rate of the ABC-SMC algorithm (the process should stop if the success rate falls below it)}
}
\description{
Run ABC-SMC with DAISIE-IW data
}
\keyword{datasets}
