% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_functions.R
\name{prior_dens_onlyABC}
\alias{prior_dens_onlyABC}
\title{prior density function}
\usage{
prior_dens_onlyABC(pars, idparsopt)
}
\value{
Density of the given parameter set
}
\description{
prior density function
}
