% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_functions.R
\name{prior_dens_spi}
\alias{prior_dens_spi}
\title{prior density function for small-phylogeny islands}
\usage{
prior_dens_spi(pars, idparsopt)
}
\value{
Density of the given parameter set
}
\description{
prior density function for small-phylogeny islands
}
