% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_functions.R
\name{prior_gen_spi}
\alias{prior_gen_spi}
\title{prior function to generate parameters}
\usage{
prior_gen_spi(pars, idparsopt)
}
\value{
a vector of parameters
}
\description{
prior function to generate parameters
}
