% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ABC_par.R
\name{run_ABC_par}
\alias{run_ABC_par}
\title{Run ABC-SMC with DAISIE-IW data}
\usage{
run_ABC_par(
  param_set,
  idparsopt,
  ss_set = 0,
  number_of_particles = 100,
  num_iterations = 20,
  print_frequency = 20,
  sigma = 0.05,
  stop_rate = 1e-05,
  saveOrNot = TRUE,
  num_threads = 1,
  start_of_file_name
)
}
\arguments{
\item{param_set}{A numeric of a specific line of parameter set in parameter space}

\item{idparsopt}{A vector of positions of the parameters that be optimized.}

\item{number_of_particles}{The number of particles in each iteration.}

\item{num_iterations}{The maximum number of iterations.}

\item{print_frequency}{The frequency of printing out the progress.}

\item{sigma}{The standard deviation of the normal distribution used to generate the proposal distribution.}

\item{stop_rate}{The stopping rate of the ABC-SMC algorithm (the process should stop if the success rate falls below it)}

\item{num_threads}{number of threads}
}
\description{
Run ABC-SMC with DAISIE-IW data
}
