% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_SMC_iw_par.R
\name{Thijs}
\alias{Thijs}
\alias{process_particle}
\title{process particles}
\usage{
process_particle(
  par_values,
  prior_density_function,
  idparsopt,
  calc_ss_function,
  obs_data,
  epsilon_values
)
}
\arguments{
\item{par_values}{a list of parameters}
}
\description{
process particles
}
\keyword{internal}
