% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_error_all.R
\name{calc_error_all}
\alias{calc_error_all}
\title{Calculate summary statistic distances between two simulated trees}
\usage{
calc_error_all(sim_1, sim_2, replicates = 1, distance_method = "abs")
}
\description{
Calculate summary statistic distances between two simulated trees
}
