% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_error_all.R
\name{calc_error_all}
\alias{calc_error_all}
\title{Calculate summary statistic distances between two simulated trees when using all statistics
Diversity difference: \code{num_nonend}, \code{num_sington}, \code{num_multi}
NLTT difference: \code{nonend_nltt}, \code{singleton_nltt}, \code{multi_nltt}
Clade size difference: \code{first_cs_diff}, \code{prop_largest_clade_diff}, \code{rank_largest_clade_diff}, \code{clade_evenness_diff}
Colonisation difference: \code{sd_colon_time}, \code{num_colon}}
\usage{
calc_error_all(sim_1, sim_2, replicates = 1, distance_method = "abs")
}
\value{
A vector with all error metrics:
}
\description{
Calculate summary statistic distances between two simulated trees when using all statistics
Diversity difference: \code{num_nonend}, \code{num_sington}, \code{num_multi}
NLTT difference: \code{nonend_nltt}, \code{singleton_nltt}, \code{multi_nltt}
Clade size difference: \code{first_cs_diff}, \code{prop_largest_clade_diff}, \code{rank_largest_clade_diff}, \code{clade_evenness_diff}
Colonisation difference: \code{sd_colon_time}, \code{num_colon}
}
